# hardware_store/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для магазина бытовой техники",
        "title": "Hardware Store API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/suppliers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Возвращает список всех поставщиков в системе",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "Получить список поставщиков",
                "responses": {
                    "200": {
                        "description": "Список поставщиков успешно получен",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/hardware_store_internal_web_dto.SupplierResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при получении списка",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Создаёт нового поставщика вместе с его адресом в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "Создать нового поставщика",
                "parameters": [
                    {
                        "description": "Данные поставщика для создания",
                        "name": "supplier",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.SupplierRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Поставщик успешно создан",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.SupplierResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибки валидации полей или некорректный формат запроса",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при сохранении поставщика",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/suppliers/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Возвращает полные данные поставщика по уникальному идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "Получить поставщика",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID поставщика",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Поставщик успешно получен",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.SupplierResponse"
                        }
                    },
                    "400": {
                        "description": "Невалидный формат UUID",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Поставщик не найден",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.NotFoundErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при получении поставщика",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Обновляет адрес поставщика по его уникальному идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "suppliers"
                ],
                "summary": "Обновить адрес поставщика",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID поставщика",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновлённые данные адреса",
                        "name": "address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.AddressRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Адрес поставщика успешно обновлён"
                    },
                    "400": {
                        "description": "Невалидный формат UUID или ошибки валидации данных",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Поставщик не найден",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.NotFoundErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при обновлении",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "OAuth2": []
                    }
                ],
                "description": "Удаляет поставщика по уникальному идентификатору UUID",
                "tags": [
                    "suppliers"
                ],
                "summary": "Удалить поставщика",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "UUID поставщика",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Поставщик успешно удалён"
                    },
                    "400": {
                        "description": "Невалидный формат UUID",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.ValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Поставщик не найден",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.NotFoundErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера при удалении",
                        "schema": {
                            "$ref": "#/definitions/hardware_store_internal_web_dto.InternalErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "hardware_store_internal_web_dto.AddressRequest": {
            "description": "Запрос на создание нового адреса с информацией о стране, городе и улице",
            "type": "object",
            "required": [
                "city",
                "country",
                "street"
            ],
            "properties": {
                "city": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2,
                    "example": "Москва"
                },
                "country": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 2,
                    "example": "Россия"
                },
                "street": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Технопарк, 15"
                }
            }
        },
...
        "hardware_store_internal_web_dto.InternalErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
...
        "hardware_store_internal_web_dto.NotFoundErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "client not found"
                }
            }
        },
...
        "hardware_store_internal_web_dto.SupplierRequest": {
            "description": "Запрос на создание нового поставщика с контактной информацией и адресом",
            "type": "object",
            "required": [
                "name",
                "phone_number"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/hardware_store_internal_web_dto.AddressRequest"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "ООО 'ТехноСнаб'"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79561234567"
                }
            }
        },
        "hardware_store_internal_web_dto.SupplierResponse": {
            "description": "Данные поставщика включая контактную информацию и ссылку на адрес",
            "type": "object",
            "properties": {
                "address_uuid": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "ООО 'ТехноСнаб'"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79561234567"
                },
                "supplier_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "hardware_store_internal_web_dto.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid uuid format"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: \"Bearer \u003cваш_токен\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://accounts.google.com/o/oauth2/v2/auth",
            "tokenUrl": "https://oauth2.googleapis.com/token",
            "scopes": {
                "email": "Просмотр вашего email",
                "profile": "Просмотр вашего профиля"
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        },
        {
            "OAuth2": []
        }
    ]
}